Created Python/Basic-Machine-Learning
Thanks to Tech with Tim for the great tutorial on YouTube!

I went through a multitude of combinations of the input attributes and it seems G2 is always the major contributor to G3, with failures decrementing the Final Grade as failures increases. Family relationship (famrel) also slightly factors in, but not enough to affect final accuracy non-negligibly.

I chose to Visually Plot The affect of each student's second-period grade, failures, daily alcohol consumption, and weekly alcohol consumption on their Final Grade

I found that if alcohol consumption of any kind increases, then the Final Grade decreases.